+++ include/InteractiveToolkit/Platform/Core/ThreadDataSet.h
@@ -73,12 +73,12 @@ namespace Platform
             register_mutex.lock();
 
             // grab all semaphores
-            semaphore.acquireCount(threadData.size());
+            semaphore.acquireCount((uint32_t)threadData.size());
 
             threadData[tid] = (Thread *)(thread);
 
             // release all semaphores
-            semaphore.releaseCount(threadData.size());
+            semaphore.releaseCount((uint32_t)threadData.size());
 
             register_mutex.unlock();
